align dtype

+++ tests/distributed/test_pipeline_parallel.py
@@ -66,7 +66,7 @@ def test_compare_tp(TP_SIZE, PP_SIZE, EAGER_MODE, CHUNKED_PREFILL,
     tp_args = [
         # use half precision for speed and memory savings in CI environment
         "--dtype",
-        "bfloat16",
+        "float16",
         "--max-model-len",
         "8192",
         "--tensor-parallel-size",
